Refactor Cell data to be in its own file
